sync_raft_topology_nodes: parallelize system_keyspace update functions
In sync_raft_topology_nodes we execute a system keyspace
update query for each node of the cluster. The system keyspace
tables use schema commitlog which by default enables use_o_dsync.
This means that each write to the commitlog is accompanied by fsync.
For large clusters this can incur hundreds of writes with fsyncs, which
is very expensive. For example, in #17039 for a  moderate size cluster
of 50 nodes sync_raft_topology_nodes took almost 5 seconds.

In this commit we solve this problem by running all such update
queries in parallel. The commitlog should batch them and issue
only one write syscall to the OS.

Closes scylladb/scylladb#17243