service: don't loose service_level_controller when bouncing client_state
When bounce_to_shard happens we need to fill client_state with
sl_controller appropriate for destination shard.

Before the patch sl_controller was set to null after the bounce.
It was fine becauase looks like it was never used in such scenario.
With auth-v2 we need to bounce attach/detach service level statements
because they modify things via auth subsystem which needs to be called
on shard 0.