tree: remove unnecessary yields around for_each_tablet()
Commit 904bafd069a5fc94e473113ff9fb8703f433a8e5 consolidated the two
existing for_each_tablet() overloads, to the one which has a future<>
returning callback. It also added yields to the bodies of said
callbacks. This is unnecessary, the loop in for_each_tablet() already
has a yield per tablet, which should be enough to prevent stalls.

This patch is a follow-up to #17118