raft topology: ip change: purge old IP
When a node changes IP address we need to
remove its old IP from system.peers and
gossiper.

We do this in sync_raft_topology_nodes when
the new IP is saved into system.peers to avoid
losing the mapping if the node crashes
between deleting and saving the new IP. In the
next commit we handle the possible duplicates
in this case by dropping them on the read path.

In subsequent commits, test_change_ip will be
adjusted to ensure that old IPs are removed.

fixes scylladb/scylladb#16886
fixes scylladb/scylladb#16691