tablets: add warning on CREATE KEYSPACE
The CDC feature is not supported on a table that uses tablets
(Refs #16317), so if a user creates a keyspace with tablets enabled
they may be surprised later (perhaps much later) when they try to enable
CDC on the table and can't.

The LWT feature always had issue Refs #5251, but it has become potentially
more common with tablets.

So it was proposed that as long as we have missing features (like CDC or
LWT), every time a keyspace is created with tablets it should output a
warning (a bona-fide CQL warning, not a log message) that some features
are missing, and if you need them you should consider re-creating the
keyspace without tablets.

This patch does this. It was surprisingly hard and ugly to find a place
in the code that can check the tablet-ness of a keyspace while it is
still being created, but I think I found a reasonable solution.

The warning text in this patch is the following (obviously, it can
be improved later, as we perhaps find more missing features):

   "Tables in this keyspace will be replicated using tablets, and will
    not support the CDC feature (issue #16317) and LWT may suffer from
    issue #5251 more often. If you want to use CDC or LWT, please drop
    this keyspace and re-create it without tablets, by adding AND TABLETS
    = {'enabled': false} to the CREATE KEYSPACE statement."

This patch also includes a test - that checks that this warning is is
indeed generated when a keyspace is created with tablets (either by default
or explicitly), and not generated if the keyspace is created without
tablets.

Obviously, this entire patch - the warning and its test - can be reverted
as soon as we support CDC (and all other features) on tablets.

Fixes #16807

Signed-off-by: Nadav Har'El <nyh@scylladb.com>