locator: topology: use frozen_sstring for dc_name and rack_name
The datacenter and rack names never change.
We can take advantage of that by keeping a frozen_sstring
that pre-calculates their hash value so they are more
efficient to use as map keys and fast to compare
(against strings of the same size which is typical
for datacenter and rack names).

The change ripples everywhere, but the change is mechanical
and the risk is low.  On the contrary, having dc_name
and rack_name be stringly typed reduces the risk of an sstring
holding a dc name used to compare against a rack name.

FWIW, the change does not affect perf_simple_query.
We probably need a test with a large number of datacenters
and racks to see any difference.

Before (f990ea9678):
median 124269.34 tps ( 61.1 allocs/op,  13.1 tasks/op,   41728 insns/op,        0 errors)

After:
median 125765.35 tps ( 61.1 allocs/op,  13.1 tasks/op,   41731 insns/op,        0 errors)

* Used `build/release/scylla perf-simple-query -c 1`
  under `sudo cpupower frequency-set -g userspace`

Refs #17100

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>