on_endpoint_change: coroutinize the lambda around sync_raft_topology_nodes
We introduce the helper 'ensure_alive' which takes a
coroutine lambda and returns a wrapper which
ensures the proper lifetime for it.
It works by moving the input lambda onto the heap and
keeping the ptr alive until the resulting future
is resolved.

We also move the holder acquired from _async_gate
to the 'then' lambda closure, since now these closures
will be kept alive during the lambda coroutine execution.

We'll be adding more code to this lambda in the subsequent
commits, it's easier to work with coroutines.