test/cql-pytest: fix guadrail tests to not be sensitive to more warnings
The guardrail tests check that certain guardrails enable and disable
certain warnings.

These tests currently check for the *number* of warnings returned by a
request, assuming that without the guardrail there would be no warning.
But in the following patch we plan to add an additional warning on
keyspace creation (that warns about tablets missing some features).
So the tests should check for whether or not a *specific* warning is
returned - not the count.

I only modified tests which the change in the next patch will break.
Tests which use SimpleStrategy and will not get the extra warning,
are unmodified and continue to use the old approach of counting
warnings.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>